backend/ml-service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.20.0
psycopg2-binary>=2.9.0
redis>=4.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=0.20.0
# sentence-transformers>=2.2.0
# nltk>=3.8.0